squash migrations
