--- pocketbase/pb_migrations/1710992733_collections_snapshot.js
@@ -1,149 +1,0 @@
-/// <reference path="../pb_data/types.d.ts" />
-migrate((db) => {
-  const snapshot = [
-    {
-      "id": "_pb_users_auth_",
-      "created": "2024-03-19 04:18:26.360Z",
-      "updated": "2024-03-21 02:56:40.017Z",
-      "name": "users",
-      "type": "auth",
-      "system": false,
-      "schema": [
-        {
-          "system": false,
-          "id": "users_name",
-          "name": "name",
-          "type": "text",
-          "required": false,
-          "presentable": false,
-          "unique": false,
-          "options": {
-            "min": null,
-            "max": null,
-            "pattern": ""
-          }
-        },
-        {
-          "system": false,
-          "id": "users_avatar",
-          "name": "avatar",
-          "type": "file",
-          "required": false,
-          "presentable": false,
-          "unique": false,
-          "options": {
-            "mimeTypes": [
-              "image/jpeg",
-              "image/png",
-              "image/svg+xml",
-              "image/gif",
-              "image/webp"
-            ],
-            "thumbs": null,
-            "maxSelect": 1,
-            "maxSize": 5242880,
-            "protected": false
-          }
-        },
-        {
-          "system": false,
-          "id": "mwn7vztx",
-          "name": "role",
-          "type": "select",
-          "required": false,
-          "presentable": false,
-          "unique": false,
-          "options": {
-            "maxSelect": 1,
-            "values": [
-              "admin",
-              "user"
-            ]
-          }
-        }
-      ],
-      "indexes": [],
-      "listRule": "id = @request.auth.id",
-      "viewRule": "id = @request.auth.id",
-      "createRule": "",
-      "updateRule": "id = @request.auth.id",
-      "deleteRule": "id = @request.auth.id",
-      "options": {
-        "allowEmailAuth": true,
-        "allowOAuth2Auth": true,
-        "allowUsernameAuth": true,
-        "exceptEmailDomains": null,
-        "manageRule": null,
-        "minPasswordLength": 8,
-        "onlyEmailDomains": null,
-        "onlyVerified": false,
-        "requireEmail": false
-      }
-    },
-    {
-      "id": "t2axrfmcwthb802",
-      "created": "2024-03-21 00:02:16.313Z",
-      "updated": "2024-03-21 03:26:21.070Z",
-      "name": "trips",
-      "type": "base",
-      "system": false,
-      "schema": [
-        {
-          "system": false,
-          "id": "xymfmozb",
-          "name": "name",
-          "type": "text",
-          "required": false,
-          "presentable": false,
-          "unique": false,
-          "options": {
-            "min": null,
-            "max": null,
-            "pattern": ""
-          }
-        },
-        {
-          "system": false,
-          "id": "ugaebmat",
-          "name": "startDate",
-          "type": "text",
-          "required": false,
-          "presentable": false,
-          "unique": false,
-          "options": {
-            "min": null,
-            "max": null,
-            "pattern": ""
-          }
-        },
-        {
-          "system": false,
-          "id": "kwvlaxze",
-          "name": "endDate",
-          "type": "text",
-          "required": false,
-          "presentable": false,
-          "unique": false,
-          "options": {
-            "min": null,
-            "max": null,
-            "pattern": ""
-          }
-        }
-      ],
-      "indexes": [],
-      "listRule": "@request.auth.id != \"\"",
-      "viewRule": "@request.auth.id != \"\"",
-      "createRule": "@request.auth.id != \"\"",
-      "updateRule": "@request.auth.role = 'admin'",
-      "deleteRule": "@request.auth.role = 'admin'",
-      "options": {}
-    }
-  ];
-
-  const collections = snapshot.map((item) => new Collection(item));
-
-  return Dao(db).importCollections(collections, true, null);
-}, (db) => {
-  return null;
-})
