--- pocketbase/pb_migrations/1711136603_created_courses.js
@@ -1,41 +1,0 @@
-/// <reference path="../pb_data/types.d.ts" />
-migrate((db) => {
-  const collection = new Collection({
-    "id": "5lbvej7sjeja9i3",
-    "created": "2024-03-22 19:43:23.255Z",
-    "updated": "2024-03-22 19:43:23.255Z",
-    "name": "courses",
-    "type": "base",
-    "system": false,
-    "schema": [
-      {
-        "system": false,
-        "id": "xinjkrru",
-        "name": "name",
-        "type": "text",
-        "required": false,
-        "presentable": false,
-        "unique": false,
-        "options": {
-          "min": null,
-          "max": null,
-          "pattern": ""
-        }
-      }
-    ],
-    "indexes": [],
-    "listRule": "",
-    "viewRule": "",
-    "createRule": "",
-    "updateRule": "",
-    "deleteRule": null,
-    "options": {}
-  });
-
-  return Dao(db).saveCollection(collection);
-}, (db) => {
-  const dao = new Dao(db);
-  const collection = dao.findCollectionByNameOrId("5lbvej7sjeja9i3");
-
-  return dao.deleteCollection(collection);
-})
