--- pocketbase/pb_migrations/1711139314_updated_teeBoxes.js
@@ -1,20 +1,0 @@
-/// <reference path="../pb_data/types.d.ts" />
-migrate((db) => {
-  const dao = new Dao(db)
-  const collection = dao.findCollectionByNameOrId("wk0de5qc1hytuou")
-
-  collection.listRule = ""
-  collection.viewRule = ""
-  collection.createRule = ""
-
-  return dao.saveCollection(collection)
-}, (db) => {
-  const dao = new Dao(db)
-  const collection = dao.findCollectionByNameOrId("wk0de5qc1hytuou")
-
-  collection.listRule = null
-  collection.viewRule = null
-  collection.createRule = null
-
-  return dao.saveCollection(collection)
-})
