--- pocketbase/pb_migrations/1711150408_updated_trips.js
@@ -1,16 +1,0 @@
-/// <reference path="../pb_data/types.d.ts" />
-migrate((db) => {
-  const dao = new Dao(db)
-  const collection = dao.findCollectionByNameOrId("t2axrfmcwthb802")
-
-  collection.listRule = null
-
-  return dao.saveCollection(collection)
-}, (db) => {
-  const dao = new Dao(db)
-  const collection = dao.findCollectionByNameOrId("t2axrfmcwthb802")
-
-  collection.listRule = "@request.auth.id != \"\""
-
-  return dao.saveCollection(collection)
-})
