--- pocketbase/pb_migrations/1711167649_updated_holeScores.js
@@ -1,24 +1,0 @@
-/// <reference path="../pb_data/types.d.ts" />
-migrate((db) => {
-  const dao = new Dao(db)
-  const collection = dao.findCollectionByNameOrId("hnh6zsmtg1u5stl")
-
-  collection.listRule = ""
-  collection.viewRule = ""
-  collection.createRule = ""
-  collection.updateRule = ""
-  collection.deleteRule = ""
-
-  return dao.saveCollection(collection)
-}, (db) => {
-  const dao = new Dao(db)
-  const collection = dao.findCollectionByNameOrId("hnh6zsmtg1u5stl")
-
-  collection.listRule = null
-  collection.viewRule = null
-  collection.createRule = null
-  collection.updateRule = null
-  collection.deleteRule = null
-
-  return dao.saveCollection(collection)
-})
